+++ src/models/message-restful-model-collection.ts
@@ -1,0 +1,45 @@
+import RestfulModelCollection from './restful-model-collection';
+import Message from './message';
+import NylasConnection from '../nylas-connection';
+
+export default class MessageRestfulModelCollection extends RestfulModelCollection<
+  Message
+> {
+  connection: NylasConnection;
+  modelClass: typeof Message;
+
+  constructor(connection: NylasConnection) {
+    super(Message, connection);
+    this.connection = connection;
+    this.modelClass = Message;
+  }
+
+  /**
+   * Return Multiple Messages by a list of Message IDs.
+   * @param messageIds The list of message ids to find.
+   * @param options Additional options including: view, offset, limit, and callback
+   * @returns The list of messages.
+   */
+  findMultiple(
+    messageIds: string[],
+    options?: {
+      view?: string;
+      offset?: number;
+      limit?: number;
+      callback?: (error: Error | null, results?: Message[]) => void;
+    }
+  ): Promise<Message[]> {
+    if (options && options.view) {
+      // view is a parameter, so move it into a params object
+      (options as Record<string, unknown>).params = {
+        view: options.view,
+      };
+      delete options.view;
+    }
+
+    return this.range({
+      path: `${this.path()}/${messageIds.join()}`,
+      ...options,
+    });
+  }
+}